제가 배워야할 정도로 코드를 잘 짜셔서 궁금한점은 좀 더 보고 여쭤보겠습니다.

+++ src/main/java/com/bside/pjt/zerobackend/mission/controller/response/DailyMissionProgressResponse.java
@@ -12,7 +12,7 @@ public final class DailyMissionProgressResponse {
 
     private final MissionResponse mission;
     private final MissionProgressResponse missionProgress;
-    private final Long daysOfProgress;
+    private final Integer daysOfProgress;
 
     public static DailyMissionProgressResponse from(final DailyMissionProgressDto dto) {
         return new DailyMissionProgressResponse(
